+++ examples/style_macro/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "style_macro"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+leptos = { git = "https://github.com/leptos-rs/leptos.git" }
+stylers = { path = "../../stylers" }
