+++ examples/style_macro/index.html
@@ -1,13 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Styler</title>
 
-    <link data-trunk rel="css" href="../main.css">
+    <link data-trunk rel="css" href="./main.css">
 </head>
+
 <body>
 </body>
+
 </html>
