+++ examples/style_macro/src/lib.rs
