+++ examples/style_sheet_macro/src/hello.css
@@ -1,0 +1,40 @@
+div {
+    border: 1px solid black;
+    margin: 25px 50px 75px 100px;
+    background-color: lightblue;
+    content: "hello";
+    font: "1.3em/1.2" Arial, Helvetica, sans-serif;
+}
+
+.two {
+    color: yellow;
+}
+
+div .one p {
+    color: blue;
+}
+
+div.one {
+    color: red;
+}
+
+div #two {
+    color: blue;
+}
+
+h2,
+a {
+    color: purple;
+}
+
+.one:hover {
+    background-color: green;
+}
+
+p:lang(it) {
+    background: yellow;
+}
+
+p::before {
+    content: raw_str("Read this: ");
+}
