+++ stylers_test/src/style_sheet_test/test.css
@@ -1,0 +1,11 @@
+.two {
+    color: yellow;
+}
+
+.two.one {
+    color: yellow;
+}
+
+.two .one {
+    color: yellow;
+}
